backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1 
PyJWT==2.8.0
requests==2.31.0
xhtml2pdf==0.2.11
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.40.0

